+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- **BREAKING** supervisor: actors discard `ValidateConfig` by default.
+
 ## [0.2.0-alpha.0]
 ### Added
 - messages: `Impossible` that cannot be constructed ([#39]).